fetch data from googlesheets funktsioon
