+++ helpers/fetch_spreadsheet_data.js
@@ -3,7 +3,7 @@ const readline = require('readline');
 const {google} = require('googleapis');
 
 
-function Fetch(FetchDataCB) {
+function Connect(FetchDataCB) {
 
     // t88kataloogiks skripti kataloog
     process.chdir(__dirname);
@@ -80,26 +80,32 @@ function Fetch(FetchDataCB) {
  * @see https://docs.google.com/spreadsheets/d/1BxiMVs0XRA5nFMdKvBdBZjgmUUqptlbs74OgvE2upms/edit
  * @param {google.auth.OAuth2} auth The authenticated Google OAuth client.
  */
-// function FetchDataFromSheet(auth) {
-//   const sheets = google.sheets({version: 'v4', auth});
-//   sheets.spreadsheets.values.get({
-//     spreadsheetId: '1J_cYJnZI41V8TGuOa8GVDjnHSD9qRmgKTJR3sd9Ff7Y',
-//     range: 'Filmid',
-//   }, (err, res) => {
-//     if (err) return console.log('The API returned an error: ' + err);
-//     const rows = res.data.values;
-//     if (rows.length) {
-//     //   console.log('id, Filmid:');
-//       // Print columns A and E, which correspond to indices 0 and 4.
-//       rows.map((row) => {
-//         console.log(`${row[0]}, ${row[4]}`);
-//       });
-//     } else {
-//       console.log('No data found.');
-//     }
-//   });
-// }
+
 
 // Fetch(FetchDataFromSheet);
 
-module.exports.Fetch = Fetch
+function Fetch(spreadsheetId, range){
+    Connect(function FetchDataFromSheet(auth) {
+        const sheets = google.sheets({version: 'v4', auth});
+        sheets.spreadsheets.values.get({
+          spreadsheetId: spreadsheetId,
+          range: range,
+        }, (err, res) => {
+          if (err) return console.log('The API returned an error: ' + err);
+          const rows = res.data.values;
+          if (rows.length) {
+          //   console.log('id, Filmid:');
+            // Print columns A and E, which correspond to indices 0 and 4.
+            rows.map((row) => {
+              console.log(`${row[0]}, ${row[4]}`);
+            });
+          } else {
+            console.log('No data found.');
+          }
+        });
+      });
+}
+
+Fetch('1J_cYJnZI41V8TGuOa8GVDjnHSD9qRmgKTJR3sd9Ff7Y', 'Filmid')
+
+module.exports.Fetch = Fetch;
